add to_data_frame
The new `to_data_frame` function reads .vtu files from a specified directory, processes the data, and saves it as a CSV file (especially for converting data from FEM thermal_mechanical ). This function utilizes `pandas` for efficient data manipulation and handling.
This helps to make it easier to illustrate and use data.